quantum_algorithms/quantum_grover.py
import cirq
import numpy as np

class GroverOptimizer:
    def __init__(self, num_qubits):
        self.num_qubits = num_qubits
        self.circuit = None
        self.qubits = [cirq.GridQubit(i, 0) for i in range(num_qubits)]

    def oracle(self, marked_state):
        """Creates an oracle that marks the target state."""
        return cirq.X.on_each(*[self.qubits[i] for i in range(self.num_qubits) 
                               if not (marked_state & (1 << i))])

    def diffusion(self):
        """Implements the diffusion operator (Grover's diffusion)."""
        # Apply Hadamard gates to all qubits
        yield cirq.H.on_each(*self.qubits)
        # Apply X gates to all qubits
        yield cirq.X.on_each(*self.qubits)
        # Apply multi-controlled Z
        yield cirq.Z(self.qubits[-1]).controlled_by(*self.qubits[:-1])
        # Apply X gates to all qubits
        yield cirq.X.on_each(*self.qubits)
        # Apply Hadamard gates to all qubits
        yield cirq.H.on_each(*self.qubits)

    def build_circuit(self, marked_state):
        """Builds the complete Grover's algorithm circuit."""
        circuit = cirq.Circuit()
        
        # Initialize in superposition
        circuit.append(cirq.H.on_each(*self.qubits))
        
        # Number of Grover iterations
        iterations = int(np.pi/4 * np.sqrt(2**self.num_qubits))
        
        for _ in range(iterations):
            # Oracle
            circuit.append(self.oracle(marked_state))
            # Diffusion operator
            circuit.append(self.diffusion())
        
        # Measurement
        circuit.append(cirq.measure(*self.qubits, key='result'))
        
        return circuit

    def optimize(self, target_state):
        """Runs Grover's algorithm to find the target state."""
        circuit = self.build_circuit(target_state)
        
        # Simulate the circuit
        simulator = cirq.Simulator()
        result = simulator.run(circuit, repetitions=100)
        
        # Get the most frequent measurement result
        measurements = result.measurements['result']
        counts = np.bincount(measurements.sum(axis=1))
        measured_state = np.argmax(counts)
        
        return {
            'measured_state': measured_state,
            'probability': counts[measured_state] / 100,
            'circuit_depth': len(circuit),
            'num_qubits': self.num_qubits
        }

    def quantum_enhanced_search(self, data, target):
        """Implements quantum-enhanced search using Grover's algorithm."""
        if not data:
            return {'error': 'Empty dataset'}
            
        # Convert target to quantum state index
        try:
            target_state = data.index(target)
        except ValueError:
            return {'error': 'Target not found in dataset'}
            
        result = self.optimize(target_state)
        
        return {
            'found_index': result['measured_state'],
            'confidence': result['probability'],
            'quantum_resources': {
                'qubits_used': result['num_qubits'],
                'circuit_depth': result['circuit_depth']
            }
        }
