requirements.txt
cirq-core>=1.1.0
numpy>=1.24.0
rdkit>=2023.3.1
pubchempy>=1.0.4
chembl_webresource_client>=0.10.8
biopython>=1.81
tensorflow>=2.12.0
torch>=2.0.0
qiskit>=0.43.0
pennylane>=0.30.0
python-dotenv>=0.19.0
streamlit>=1.24.0
requests>=2.31.0
scipy>=1.11.0
